Pikuniku Robot Dance Battle Macro
If you get enough to pass but less than 100% the game has no retry option for the achievement. You have to redo the story or know to force quit reload the game as soon as you mess up.  Not fun, so this is a robot script to battle the robot.